feat: ajust authentication flow and add middlewares

+++ src/use-cases/user/generate-2fa-qrcode-key/generate-2fa-qrcode-key-usecase.ts
@@ -1,3 +1,4 @@
+import bcrypt from 'bcryptjs';
 import { inject, injectable } from 'inversify';
 import qrcode from 'qrcode';
 
@@ -28,18 +29,22 @@ export class Generate2FAQrCodeKeyUseCase implements IGenerate2FAQrCodeKeyUseCase
     @inject(TYPES.OneTimePasswordProvider) private readonly oneTimePasswordProvider: IOneTimePasswordProvider,
   ) {}
 
-  async execute({ userId }: IGenerate2FAQrCodeKeyRequest): Promise<IGenerate2FAQrCodeKeyResponse> {
-    const user = await this.userRepository.findById(userId);
+  async execute({ email, password }: IGenerate2FAQrCodeKeyRequest): Promise<IGenerate2FAQrCodeKeyResponse> {
+    const user = await this.userRepository.findByEmail(email);
 
     if (!user) throw new BusinessError(BusinessErrorCodes.USER_NOT_FOUND);
 
-    await this.userSecondFactorKeyRepository.removeUnvalidatedKeys(userId);
+    const passwordMatch = await bcrypt.compare(password, user.password);
+
+    if (!passwordMatch) throw new BusinessError(BusinessErrorCodes.INVALID_PASSWORD);
+
+    await this.userSecondFactorKeyRepository.removeUnvalidatedKeys(user.id);
 
     const key = this.oneTimePasswordProvider.generateBase32Key();
 
-    await this.userSecondFactorKeyRepository.generate(userId, key);
+    await this.userSecondFactorKeyRepository.generate(user.id, key);
 
-    const fileName = `${userId}.png`;
+    const fileName = `${user.id}.png`;
     const fileDirTmp = `${upload.tmpFolder}/${fileName}`;
     await deleteFile(fileName);
     const keyName = 'Node 2FA';
@@ -57,7 +62,7 @@ export class Generate2FAQrCodeKeyUseCase implements IGenerate2FAQrCodeKeyUseCase
       throw new BusinessError(BusinessErrorCodes.QRCODE_NOT_FOUND);
     }
 
-    const qrCodeUrl = `${ConstantsEnv.apiUrl}/users/qrcode/${userId}.png`;
+    const qrCodeUrl = `${ConstantsEnv.apiUrl}/users/qrcode/${user.id}.png`;
 
     return { qrCodeUrl };
   }
